requirements.txt
fastapi==0.62.0
fastapi-utils==0.2.1
uvicorn==0.12.3
gunicorn==20.0.4
httptools==0.1.1
requests==2.24.0
PyJWT==1.4.2
python-json-logger==0.1.11
paramiko==2.7.2
uvloop==0.16.0
